# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hospital Management System API",
        "description": "API for managing hospital operations with role-based access control",
        "contact": {
            "email": "admin@hospital.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/doctor/records": {
            "get": {
                "tags": [
                    "👩‍⚕️ Doctor Services"
                ],
                "summary": "Get all medical records",
                "description": "Retrieves all medical records created by the authenticated doctor",
                "operationId": "getAllMedicalRecords",
                "responses": {
                    "200": {
                        "description": "Medical records retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Medical records retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "medical_records": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "registration_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "diagnosis": {
                                                                "type": "string",
                                                                "example": "Hypertension"
                                                            },
                                                            "treatment": {
                                                                "type": "string",
                                                                "example": "Lifestyle modification and medication"
                                                            },
                                                            "additional_notes": {
                                                                "type": "string",
                                                                "example": "Patient should monitor blood pressure daily"
                                                            },
                                                            "input_date": {
                                                                "type": "string",
                                                                "format": "datetime"
                                                            },
                                                            "registration": {
                                                                "type": "object"
                                                            },
                                                            "medicines": {
                                                                "type": "array",
                                                                "items": {
                                                                    "type": "object"
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Doctor profile not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Doctor profile not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "👩‍⚕️ Doctor Services"
                ],
                "summary": "Create a new medical record",
                "description": "Creates a new medical record for a patient registration with optional medicines",
                "operationId": "createMedicalRecord",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "registration_id",
                                    "diagnosis",
                                    "treatment"
                                ],
                                "properties": {
                                    "registration_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "diagnosis": {
                                        "type": "string",
                                        "example": "Hypertension Grade 1"
                                    },
                                    "treatment": {
                                        "type": "string",
                                        "example": "Lifestyle modification and antihypertensive medication"
                                    },
                                    "additional_notes": {
                                        "type": "string",
                                        "example": "Patient should monitor blood pressure daily and return in 2 weeks"
                                    },
                                    "medicines": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "medicine_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 30
                                                },
                                                "usage_instructions": {
                                                    "type": "string",
                                                    "example": "Take 1 tablet daily after breakfast"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Medical record created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Medical record created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "medical_record": {
                                                    "type": "object"
                                                },
                                                "medicines": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized to create record for this registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized to create medical record for this registration"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or insufficient medicine stock",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/patient/medical-records": {
            "get": {
                "tags": [
                    "🏥 Patient Services"
                ],
                "summary": "Get patient's own medical records",
                "description": "Retrieves all medical records for the authenticated patient",
                "operationId": "getPatientMedicalRecords",
                "responses": {
                    "200": {
                        "description": "Medical records retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Medical records retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "medical_records": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/MedicalRecordResource"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Patient profile not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Patient profile not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MedicalRecordResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "registration_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "diagnosis": {
                        "type": "string",
                        "example": "Hypertension Grade 1"
                    },
                    "treatment": {
                        "type": "string",
                        "example": "Lifestyle modification and medication"
                    },
                    "additional_notes": {
                        "type": "string",
                        "example": "Patient should monitor blood pressure daily"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}